Updated openstack/openstack Project: openstack/nova  4e53a1cbe810a447af96a1c36e558aecf99c156c
Add functional test for bad res class in set_inventory_for_provider

The unit tests covering the case that an invalid resource class
name was passed to set_inventory_for_provider were a bit too
unit-y and we can get the same coverage down the full stack with
a simple addition to the existing functional tests for the report
client. This adds the simple change to the functional test and removes
the now redundant unit tests.

Change-Id: Idb579acf6852650d4d5a6bc7704f123a40258f2d