Updated openstack/openstack Project: openstack/neutron  ff3446796cd512c75690d3f29da813e98ffa5ebe
Fixes crash when starting Neutron OVS Agent on Windows

The neutron OVS Agent does not start on Windows because
of Linux specific import.

The fix consists in moving the method get_ip_version in
the common module of ip_lib being the only one used from
the module. We will import the common module
agent/common/ip_lib after that.

Change-Id: I95e9a6fa99e0992499f46a4f0debccd47b8a936c
Closes-Bug: #1670402