Updated openstack/openstack Project: openstack/neutron  72916b54c97db60e8cf4a3fe56e5930cfd5aa975
Get rid of delete_subnet method in ML2

This method has grown to be one of the most hideous methods in
Neutron. It's well past Icehouse so it's time to address the
REVISIT and get rid of it entirely. :)

This patch refactors ML2 and the db_base_plugin_v2 delete_subnet
method so ML2 can perform its ML2-specific tasks with callbacks
and db_base_plugin_v2 can handle the rest.

Partial-Bug: #1666424
Change-Id: Ie29acfbe323b60205ade9d660f7497f5bf4a35ca