Updated openstack/openstack Project: openstack/nova  fa2b4a82648101826566da68dd56d204e269853f
Refactor cell loading in compute/api

This merely pulls out the loading of the global cells cache
into a helper function that can be used by later patches that
will also depend on this. Since it's the same code, existing
tests exercise this path just like they did before it was moved.

This was moved out of another patch to re-order some dependencies.

Related to blueprint cells-aware-api
Change-Id: Idb25f5905f3e33cec3c7887d08408251fef5b262