Updated openstack/openstack Project: openstack/tripleo-heat-templates  09af40f5dbbdd335005640b538ed043dda75e77b
Set number of Swift proxy server workers to auto

Setting the default Swift proxy workers to 0 actually results in a
single Swift proxy worker, no matter how many CPU cores are available.
This is not the default Swift setting and is most likely not sufficient.

Setting this to auto uses the default in Swift, which equals to the
number of CPU cores.

Closes-Bug: 1655070
Change-Id: Ic321b6111f8697ba3cc1554611fee44c2e540759