Updated openstack/openstack Project: openstack-dev/devstack  e4b224928fb3bae1bbb086f429dfd3018375ace6
Adding haproxy package to Neutron

This patch adds haproxy package to devstack as Neutron will rely on it
for serving metadata instead of the current Python implementation.

haproxy will reduce the memory footprint from ~50MB to ~1.5MB for
serving metadata. It will be spawned for every Neutron router so,
for large deployments, it will be a significant memory reduction.

Change-Id: I36a5531cacc21c0d4bb7f20d4bec6da65d04c262