Updated openstack/openstack Project: openstack/horizon  03a81c54d056b2e42105a2d252bc98ed3aff311b
Refactor Project Volumes stand-alone panel

Note that there was state leaking from one of the existing
tests that was relied upon in other tests and moving that test
in the run order caused the others to fail. All related tests
have been altered to not leak state.

Change-Id: I972bc5650fa77044de8a027f570cf2cb41febef8
Implements: blueprint reorganise-volumes