Updated openstack/openstack Project: openstack/cinder  3c2fe88af0c8a1db017847cff3cea4c532f3e806
api-ref: v2 API does not mention GET volume metadata by key

In v2:
Add an explanation regarding GET for volume metadata for a specific key
and some parameters regarding key, because those are missed.
Modify prameters of volume_id and key for DELETE, because those should be in path.
In v3:
Add explanations regarding GET and DELETE for volume metadata for a specific key
and some parameters regarding key, because those are missed.

Change-Id: I0a9dd9b62bb196db97af80dc928501ad7c57df6e
Closes-Bug: #1669142