Updated openstack/openstack Project: openstack/nova  54629ff9440740d352534e53f4af727c7750a14f
Use Cinder API v3 as default

This patch sets Nova to use the Cinder v3 API as the default option.
This change uses the base '3.0' version, which is identical to v2.

Leveraging microversions that were introduced in v3 and switching to
a newer microversion will be added in a follow up patch.

Change-Id: Ibcc8eefcc5011ad5002d608a3c253f1e9b68416f