Updated openstack/openstack Project: openstack/tripleo-heat-templates  90a0c87608e8f9fe37acd659ac6b130791a743e8
Keystone token flush cron job should log to a file

Keystone UUID tokens require the token_flush job to delete expired
tokens to prevent the disk from filling.  When the job runs, it should
be allowed to log to the disk so that the job can be traced if required.

Change-Id: I62e36e0968902564b97093a45df15e963ad08242
Closes-Bug: #1648174