Updated openstack/openstack Project: openstack/os-collect-config  9d23b907fa73a0b0332758ab17fee8d254e69eac
Add Constraints support

Adding constraints support to libraries is slightly more complex than
services as the libraries themselves are listed in upper-constraints.txt
which leads to errors that you can't install a specific version and a
constrained version.

This change adds constraints support by also adding a helper script to
edit the constraints to remove os-collect-config.

Change-Id: I2a0983f41372b52e6d4e6a85cca1c51f980af36c