Updated openstack/openstack Project: openstack/os-cloud-config  0eb6aa6d99fe06c6c44f06d35a3274f070ebffd5
Add Constraints support

Adding constraints support to libraries is slightly more complex than
services as the libraries themselves are listed in upper-constraints.txt
which leads to errors that you can't install a specific version and a
constrained version.

This change adds constraints support by also adding a helper script to
edit the constraints to remove os-cloud-config.

Change-Id: Ief6350a34c682635c1edc8eca79d28fe53aff9fb