Updated openstack/openstack Project: openstack/cinder  83b137657a5d9c8f3c416aba9ccb6b72208bc09c
Set backup available after verify

In backup import request, the backup is set to available before
it is verified. Then if the verification failed, the backup's
status is set to error. This leads to an unreasonble status change:
creating -> available -> error.

The available status should be set after the verification when
all work is done. And the status change should be like:
creating -> available, or creating -> error

Closes-bug: #1669628
Change-Id: I67f30df8a8112c953f24f6d2f7d5152416d4f0df