Updated openstack/openstack Project: openstack/python-neutronclient  aaa37b62e88e0538b7f17d936f33507e6dd143ae
Don't sort the fw_rule order in OSC

Currently, the FWaaS accept the "fw_rule_list" argrument in fw_policy
update, and it will update all the related fw_rules association. If the
order is different, the behavior is not expected, so we should take care
of the input value order, and maintain the original one.

This patch remove the "sorted" function when construct the firewall policy
request body, and follow the input order.

Closes-Bug: #1671338
Change-Id: Iaa45446eee199582f5f11a4705e82fb0ab644bcc