Updated openstack/openstack Project: openstack-infra/project-config  0b4447034d5f94ba723c79485baed82ac6a38391
Enable DEVSTACK_GATE_TEMPEST_NOTESTS for patrole.

This change is the last needed to correctly execute patrole
tempest jobs. This flag needs to be set to 1, so that no
automatic tempest tests are executed in devstack-vm-gate.sh
because they are instead executed in post_test_hook.sh in
patrole's repo.

Change-Id: Id74a10d91b05090d27a5dedd73de65efbd66ec44