Updated openstack/openstack Project: openstack/tempest  9f81613888e18399e215f61258a5ccfca277402d
Add a bit more debugging to the client registry

The service client registry raises an exception in case of
duplicate registration, but it logs absolutely nothing in case
of successful one, which makes life difficult when debugging
issues with plugin registrations.

Add a debug line for the success case as well, which includes
the plugin name and the configuration data that was loaded.

Change-Id: Ic2036e9e2db4b1604a397baa0cfdaee665a477f8