Updated openstack/openstack Project: openstack/ceilometer  8c81d02efc4f436f3ebb8e15e0311fb044a8c42f
Bugfix: use transport_url from [oslo_messaging_notifications] if present

According to https://docs.openstack.org/developer/oslo.messaging/opts.html#oslo-messaging-notifications
a transport_url defined in [oslo_messaging_notifications] should be used, if
present. In Ceilometer, this is ignored in favour of the transport_url defined
in [DEFAULT], which is a bug. This change fixes the bug.

Change-Id: I12a559ebbbc1edd18c6833943bb5e9af7adf4666