Updated openstack/openstack Project: openstack-infra/storyboard  129709379284bcc687375c0f248da3259bba0b56
Add a UUID to the temporary email address

This fixes a bug where two users from Launchpad could
end up with the same temporary email address in StoryBoard,
if they both have the same name. This would cause the
migration of projects to fail since we require email
addresses to be unique.

Change-Id: I7f8a75b9702d81f336cfcd6881ac2d6614ee992c