Updated openstack/openstack Project: openstack/nova  0f29badb779df8367f0dec2f4f9880a221babaf1
No API cell up-call to delete consoleauth tokens

One goal of Cells v2 is to remove all up-calls from compute to API
cells.  This change removes the up-call to delete consoleauth tokens
during instance deletion and live migration.  Instead, the tokens are
deleted near the start of these processes in the API cell.  To close
the loop, new consoleauth tokens are not granted to instances in
deleting or migrating task states.

Change-Id: Ic7c68b75e22bad2b0237a2f861254611a86516f3