Updated openstack/openstack Project: openstack/horizon  4b7facc707c58bd17bf9eb3ba3bafc73718622a6
Provide available physical networks on create network modal

Added setting physical_networks to provide users list of
available physical networks to choose from on the
admin create network modal.

Default to an empty list and the physical network field is
a regular input field where users can type in the name
of the physical network to be used.

If it's set to a list of physical network names, the
physical network field is a dropdown menu where users can
choose a physical network to use.

Closes-bug: #1485144

Change-Id: I342b3776a52179d5b4a704fb984912878ff3dc81