Updated openstack/openstack Project: openstack/zaqar  0dc086828a22b17be401a23ca6f658fba2036538
Introduce Guru to Zaqar

Guru is a mechanism whereby developers
and system administrators can generate
a report about the state of a running
Zaqar executable.

This report is called a *Guru Meditation Report*

This mechanism will help developer or operator
to fix issues in (production) deployments without
stopping Zaqar service.

Implements: blueprint introduce-guru-to-zaqar
Change-Id: I72885be396be7eb0a9dd8fd564d706a8351b02c6