Updated openstack/openstack Project: openstack/diskimage-builder  8b19f7f99d18c22ea5aa64bf45075e673503b360
Fix dib-init-system for Debian Jessie and Debian Stretch

On Debian Jessie and Debian Stretch systemctl is in /bin.
If the package systemd-sysv is not installed the script
dib-init-system did not find the init system.
This patch fixes the problem: it also looks in /bin
for systemctl and if found decides for systemd.

Change-Id: I5a18052a070bad5e16b14672237a1e2b38513949
Signed-off-by: Andreas Florath <andreas@florath.net>