Updated openstack/openstack Project: openstack-infra/shade  98d78c5d1111aea68971ccabd46e813acd3b9b24
Only do fnmatch compilation and logging once per loop

We don't need to compile the fnmatch regex in every loop pass - so
compile it at the top of the loop. Also, logging about a bad regex on
every loop is silly. Only log if we don't have a match at the end.

Change-Id: Iffa039237042d6a2f75aa08f2c949d508b367ec4