Updated openstack/openstack Project: openstack-infra/project-config  117f0a99c33865dab6d50167cbda42369fbef9b4
Add os_molteniron to OpenStack-Ansible

This patch adds new pre-seeded git repository to OpenStack-Ansible for
deploying Molteniron.

This project will have its own core team with the general
openstack-ansible core team included, with the appropriate rights to
ensure that they can act as a back-stop in case of any issues.

Needed-By: Ie7190e6d46091ff91b108baf48ecf5d304885124
Change-Id: Ib60593085179819e566068576be08ee403221d22