Updated openstack/openstack Project: openstack-infra/project-config  5accc0fb4c21e4c59afc10c29b76ddc320fa9fb8
Reload sshd on centos for allow-local-ssh-root

We are seeing some issues restarting SSHD for centos.  Because we are
doing config file changes, and SSHD is running, we can use the reload
command.

Reload is a little different then restart, specifically reload does a
SIGUP on the process. In some local testing, this seems to be more
reliable to restart, which causes some bind address errors.

Change-Id: Ic5e00e1e1eed1de87c430180a6f00d30410a95b4
Signed-off-by: Paul Belanger <pabelanger@redhat.com>