Updated openstack/openstack Project: openstack-infra/project-config  4b982f7d1e992a71b906157e6ede99060725d27d
Add a new scenario only multinode gate job

This job is an attempt to reduce the load in the SUT in the gate,
by splitting API and tempest tests in two separate jobs, and by
running scenario ("heavier") tests in a dedicate job serially.

We already stopped running a number of scenario tests from the
main gate so we need this job at least for tempest to make sure
other scenario tests are not bit rotting.

This job is non voting for now.

Change-Id: I11eb21686351a98e55fecc15a794397da6d92b40