Updated openstack/openstack Project: openstack/tempest  2536646de5f739b709324a6516eb02efaa578535
Use tempest.lib data_utils - scenario

Tempest has duplicated modules of data_utils, one is tempest.lib and
the other is tempest.common. The difference is tempest.common module
refers to the config option resources_prefix. And the option is marked
as deprecated already. So it is nice to replace the calls with tempest.lib
one.

Change-Id: Ifffa05bb80b0d09b19bf2ecca0009c27642b9959