Updated openstack/openstack Project: openstack/tempest  b960e490dbb673edb63180997b1bcdab96ddba69
Use tempest.lib data_utils - compute

Tempest has duplicated modules of data_utils, one is tempest.lib and
the other is tempest.common. The difference is tempest.common module
refers to the config option resources_prefix. And the option is marked
as deprecated already. So it is nice to replace the calls with tempest.lib
one.

Change-Id: Idfa9824d14a1e5e4291fcb70bb2c00caf9083d67