Updated openstack/openstack Project: openstack/tempest  32cac095fa1a434509c4cd62c77830b063f67a0d
Use tempest.lib data_utils - identity

Tempest has duplicated modules of data_utils, one is tempest.lib and
the other is tempest.common. The difference is tempest.common module
refers to the config option resources_prefix. And the option is marked
as deprecated already. So it is nice to replace the calls with tempest.lib
one.

Change-Id: Ibd98d212ae4fc2d581ad944cc004addee77e4281