Updated openstack/openstack Project: openstack/tempest  e8f27043c9a70d478e3c93483044da7f870c2635
Removing sahara from the service available group

This is a leftover from when sahara was moved out of tree.
It causes conflicts when the sahara plugin loads, and it should
be removed.

This option is not used right now by Tempest code, so even if someone
is using it, it has not effect whatsoever on test execution.
The only case in which this option may have effect, is when the
sahara-tests Tempest plugin is installed. When the
sahara plugin is loaded, the config option will be there.

I think in this case we should skip deprecation, and just remove it
like we did in the past when moving Tempest code to a plugin.

Change-Id: Ie7556b97237b0101000d140d351edea5e518783f