Updated openstack/openstack Project: openstack/tempest  09197d458918fac8b8631776bbb8727779d057a0
Use tempest.lib data_utils - orchestration

Tempest has duplicated modules of data_utils, one is tempest.lib and
the other is tempest.common. The difference is tempest.common module
refers to the config option resources_prefix. And the option is marked
as deprecated already. So it is nice to replace the calls with tempest.lib
one.

Change-Id: I5db7cead61e5891a8e99b608069c2474ca1777a1