Updated openstack/openstack Project: openstack/heat  e32c74a1df60a4aff8656eac6b2794f5c1bc8abf
Sleep between check_resource calls

Instead of calling check_resource on all leaves in the resource graph at
once, sleep a little bit between each call. As it's a tad slower,
delegate it to a thread so that the stack_create RPC message doesn't
timeout when you have lots of resources.

Change-Id: I84d2b34d65b3ce7d8d858de106dac531aff509b7
Partial-Bug: #1566845