Updated openstack/openstack Project: openstack/horizon  9dee6b3af303331ae45007c51c24d63bd67b67c2
NG details view route should not be '/project/...'

Right now the route includes the name 'project'
but we may have details view for 'identity' and
'admin' too, so it should be more general.

Picked just 'ngdetails' instead and moved this out as a
constant. Since this constant needs to be used
by the config blocks, I moved it out into its own
constant module.

Change-Id: I7603250dd70eb40568aa74be2ae4821ee8fcefcc
Closes-Bug: #1641250