Updated openstack/openstack Project: openstack/heat-translator  d7d6aee6fcecd3736fed72caebdff1eddb99d551
Fix bug that only 1 volume will be attached to VDU automatically

With Tacker when more than 1 volume to be created and attached,
but after heat translation, only 1 volume will be attached automatically.
The code change is to allow all volumes to be created and attached as
what specified in TOSCA template.

Change-Id: Iba9e089347d6dfd31e7d7c40b46706b75145c696
Closes-Bug: #1675715