Updated openstack/openstack Project: openstack/tripleo-heat-templates  ee09ed67b9d9e021f50585543c2a407ac59a033a
keystone/containers: Add support for fernet keys

Since the 'file' resource is included in the tags that puppet takes into
account, we already generate the fernet keys if it's enabled as a token
provider.

This merely adds the keys to the container. However, if fernet is not
the provider, we make this file addition optional.

Change-Id: Id92039b3bad9ecda169323e01de7bebae70f2ba0