Updated openstack/openstack Project: openstack/tripleo-heat-templates  4a5569120175695bf8a66866c65b0850f6821098
Tasks hook for preparing BM host for deploying containerized services

This implements a host_prep_tasks hook where we can specify Ansible
tasks to perform on the host before deploying containerized
services. The hook runs in a single step, the assumption is that we will
mostly use the hook for creating per-service directories on the host to
ensure we are able to mount them into the containers. (We cannot do this
operation via Puppet because all containerized services run their Puppet
within a config container, so Puppet doesn't have access to host's
filesystem.)

Change-Id: I7d8bac39e0cd422fd651eefe29f7d10941ab4a1a