Updated openstack/openstack Project: openstack/ironic-specs  6c8c98679ac42cb60dd3542af024aab08ac2f620
Follow-up patch for the Redfish spec

This patch contains changes to the driver's properties from the original
Redfish spec[0].

Before the driver was expecting a single "redfish_uri" property which
contained two pieces of information: The URI to the root resource and
the system resource identifier.

As discussed on IRC on 14/03/2017 [1] and later on 21/03/2017 [2], it
turns out that the Redfish specification isn't very opinionated regarding
the names of the resources which may lead to some implementations to
call /Systems (as we call it in this specification) /Nodes or any other
variant. Therefore, splitting the old proposed "redfish_uri" into
"redfish_address" and "redfish_system_id" makes more sense and makes the
driver more flexible when working with different types of Redfish
systems from different vendors.

[0] 88b55d4307cdc9e6ff37d8cbfea228c1c281e79f
[1] http://eavesdrop.openstack.org/irclogs/%23openstack-ironic/%23openstack-ironic.2017-03-14.log.html#t2017-03-14T12:05:46
[2] http://eavesdrop.openstack.org/irclogs/%23openstack-ironic/%23openstack-ironic.2017-03-21.log.html#t2017-03-21T16:36:08

Change-Id: I9cdeee2f50859d97e8ae8b8b4d528446e106eced