Updated openstack/openstack Project: openstack-infra/project-config  1965ba50e5db0a650c7f0dbb0ad9e8a373c4d96b
Run 'enable_plugin' in project-config for grenade jobs

IRONIC_DISABLE_DEVSTACK_PLUGIN was added some time back to grenade jobs
in project-config to disable devstack plugin because the grenade code in
Ironic already enabled it.

Unfortunately, subnode local_conf doesn't get updated when
'enable_plugin' is in the grenade settings file. We need to move
enablement of devstack plugin back to project-config to fix this issue,
which is accomplished by this patch.

To prevent the gate from breaking due to devstack plugin being enabled
twice, a change was made to devstack/upgrade/settings that
performs a check to see if the ironic devstack plugin has already been
enabled and does the following:
  * If it has already been enabled it will not enable it again.
  * If it has not yet been enabled it will enable it.

This change was made to master, stable/ocata, and stable/newton branches.
All changes have landed.
https://review.openstack.org/#/c/444335/
https://review.openstack.org/#/c/444944/
https://review.openstack.org/#/c/444950/

Once this patch merges, we will propose a follow-up patch to
devstack/upgrade/settings that removes all the additonal logic and
enabling of the ironic devstack plugin.

Change-Id: Idb9c556e7d2553c9f13e07077676fd7505bfb19c