Updated openstack/openstack Project: openstack/neutron  1773c62525e39112a97b0cd89fb98c0e45752c7f
Untangle WaitTimeout class from eventlet.TimeoutError

This simplifies test code a bit. Previously, we introduced the
WaitTimeout class to replace the eventlet.TimeoutError that inherited
from BaseException and hence could interrupt normal test worker run.

Now that we were initiating deprecation warning messages for a cycle
when the user relied on eventlet.TimeoutError, we should be ok to remove
the parent class from the list of inheritance for WaitTimeout.

Change-Id: I6160cc1408f72a9f847a59702bc74b8e52a10497